Updated deploy.sh to take ip address as parameter

+++ deploy.sh
@@ -1,10 +1,11 @@
 #!/bin/bash
+if [ -z "$1" ]; then echo "Empty parameter for testenv IP Address"; exit -1; fi
 echo Pushing dockerfile
 
 docker push ggjonsson/tictactoe
 
 echo Executing remote...
-ssh vagrant@192.168.33.10 "docker stop tictactoe;
+ssh vagrant@"$1" "docker stop tictactoe;
 							docker rm tictactoe;
 							docker pull ggjonsson/tictactoe;
 							docker run -p 9000:8080 --name tictactoe -d -e "NODE_ENV=production" ggjonsson/tictactoe"
